requirements.txt
gurobipy==12.0.3
openpyxl==3.1.5
pandas==2.2.3
pyprojroot==0.3.0
pypsa==1.0.0
PyYAML==6.0.3
